Add PlayerLight
