--- Assets/Domain/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d24625e776382464998eda614f74aee1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
