+++ Assets/Scenes/MainBuildingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a20f518b1bf229142b0dd3899aa6f82d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
